+++ cogant/_rnd/sweep_2026_04/blocks/fixpoint_translation_api__00.py
@@ -1,0 +1,7 @@
+from cogant.translate.engine import TranslationEngine
+
+engine = TranslationEngine(max_iterations=10)
+engine.register_rule(my_rule)
+
+# Run fixpoint iteration
+mappings = engine.translate(graph)